+++ archer/src/main.rs
@@ -11,8 +11,8 @@ fn do_it(camera: &Camera, x: u32, y: u32) -> Ray {
 }
 
 fn main() {
-    const RESOLUTION: [usize; 2] = [1920, 1080];
-    let mut image: RgbImage = ImageBuffer::new(RESOLUTION[0] as u32, RESOLUTION[1] as u32);
+    const RESOLUTION: [u32; 2] = [1920, 1080];
+    let mut image: RgbImage = ImageBuffer::new(RESOLUTION[0], RESOLUTION[1]);
 
     let mut cam = Camera::new(RESOLUTION);
     cam.update();
@@ -20,8 +20,8 @@ fn main() {
     
     let now = Instant::now();
 
-    for x in 0..RESOLUTION[0] as u32 {
-        for y in 0..RESOLUTION[1] as u32 {
+    for x in 0..RESOLUTION[0] {
+        for y in 0..RESOLUTION[1] {
             let _r: Ray = do_it(&cam, x, y);      
             *image.get_pixel_mut(x, y) = image::Rgb([255; 3]);
         }
